refactor: 메소드와 클래스명 변경
