+++ src/main/java/lotto/domain/Lotteries.java
@@ -4,14 +4,18 @@ import java.util.ArrayList;
 import java.util.List;
 
 public class Lotteries {
+	private final LottoNumbersGenerator lottoNumbersGenerator = new LottoNumbersGenerator();
 	private final List<Lotto> lotteries = new ArrayList<>();
 	private final List<List<Integer>> totalLottoNumbers = new ArrayList<>();
 
 	public Lotteries(int numberOfTickets) {
-		LottoGenerator lottoGenerator = new LottoGenerator();
+		init(numberOfTickets);
+	}
+	private void init(int numberOfTickets) {
 		while (lotteries.size() != numberOfTickets) {
-			lotteries.add(new Lotto(lottoGenerator.getLotto()));
+			lotteries.add(new Lotto(lottoNumbersGenerator.getLottoNumbers()));
 		}
+		lotteries.forEach(lotto -> totalLottoNumbers.add(lotto.getNumbers()));
 	}
 
 	public List<Lotto> getLotteries() {
@@ -19,7 +23,6 @@ public class Lotteries {
 	}
 
 	public List<List<Integer>> getTotalLottoNumbers() {
-		lotteries.forEach(lotto -> totalLottoNumbers.add(lotto.getNumbers()));
 		return totalLottoNumbers;
 	}
 }
